requirements.txt
pillow
scipy
scikit-image
torchvision
tqdm
